client/requirements.txt
pywebview
pywin32